Fix #6: set process column width
`grunt-tsc` explicitly looks up the column width of the window, and then
does math with that number. In particular, it tries to trim down an
error message by a certain number of characters, but because our
window's column width is 0, it never trims the error message, and
`grunt-tsc` hangs into a while loop.

So, we use `set-process-window-size` to set our column size so that
`grunt-tsc` can pick it up properly.