Bump version to 1.0.0
Ideally, the major version bump is enough to indicate the change to the
`grunt --help` version of this module. Although the interface is still
the same ( invoke `grunt-exec` ), the experience is significantly
different ( slow delay on first call per project, prefixed version of
`grunt-exec` ) and restoring the old functionality requires end user
customization. For those two reasons, we're bumping the major version.