Improve grunt--get-help user experience
`progress-reporter` is a nifty built in that does the spinny cursor, but
it requires us to use `start-process`, which then means we need to use
`set-process-filter` and `set-process-sentinel` and
`accept-process-output` to handle the output timing.

Currently there is a `(sit-for 1)` kludge after the process completes
because there's a race condition - the process filter doesn't append the
rest of the process output to `grunt--help-string` before
`(grunt--get-help)` returns it to its caller, meaning that the output is
truncated and preventing us from parsing out the point at the end of the
tasks as regexp matched by "^$".