Refactored to make button only on required area
Since we were previously starting, and making buttons, from the start of
the buffer each time, we would be duplicating effort and repeating the
buttons each time.

This way we calculate the point of the last button created, and continue
to make buttons from this point onwards. This _may_ not be faster, since
we do the same amount of searches each time.