Added test to check grunt-task-links get reset
Another asynchronous test, we have to set up a sandbox twice as it gets
called anonymously, but we can check that after the first pass, the
variable is non nil, and after the second run _(stubbing out the
functions that change the variable)_ we assert that it gets reset.

It may be cleaner to extract this as a function which we spy on instead?
But it tests it correctly